benches before:

```
test bench_decode_1000_values        ... bench:      46,870 ns/iter (+/- 4,621)
test bench_decode_nested_empty_lists ... bench:         381 ns/iter (+/- 59)
test bench_decode_u256_value         ... bench:          54 ns/iter (+/- 9)
test bench_decode_u64_value          ... bench:          71 ns/iter (+/- 11)
```

benches after:

```
test bench_decode_1000_values        ... bench:      39,273 ns/iter (+/- 3,948)
test bench_decode_nested_empty_lists ... bench:         381 ns/iter (+/- 49)
test bench_decode_u256_value         ... bench:          50 ns/iter (+/- 9)
test bench_decode_u64_value          ... bench:          58 ns/iter (+/- 6)
```